Add scripts/xact_finisher.pl.
xact_finisher.pl is a perl script to set the xact_finish on "hanging"
circulations. There was a bug that did not set xact_finish properly on
circulations when a copy was lost and/or billings were modified. This
script is intended to fix these by setting the xact_finish on their
circulations.

It takes a patron and a copy barcode as arguments and set the xact_finish
on "open" circulations on that combination.

It currently does not do any safety checks. It assumes that the user knows
what they are doing.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>