Fix a serious bug in the previous commit.
I practically doubled the nubmer of child processes that could
be started for reingest but failed to adjust the while loop test
conditions or the number of lists that was to be checked against.
This meant that $count could to $lists before all of the lists
had been processed.

This commit, hopefully, fixes that.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>