Add active field to hold and cric matrix output in parameters.pl
This involves more than just pulling the active column in the query.
We also add a gray background to rows where the active flag is false.
This format is created in a global variable and applied when as
needed in the write_rows subroutine.

This commit also moves the bold format used in headers to a global
variable.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>