Anothter custom hack to db_upgrade.pl.
We've added an upgrade script out of order to production, and it fails
if it runs again.  We modify db_upgrade.pl to run scripts since 0949
that are missing in our database.

A variation of this hack may make it into the master branch, but it
will likely require a command line option to get the starting point
from the user.  That change will also likely cause me to use
Getopt::Long with this program.

Signed-off-by: Jason Stephenson <jstephenson@mvlcstaff.org>