Add connection options to parameters.pl
Add command line options to parameters.pl to specify the database to
connect to.  The location of the output spreadsheet also becomes a
parameter.  The following command line parameters which take arguments
were addeded:

--user: the database username
--password: the database user's password
--host: the hostname of the PostgreSQL database server
--port: the port that the PostgreSQL database server listens on
--db: the name of the PostgreSQL database
--file: filename for the output spreadsheet

The database connection parameters all have default values that make
sense for a typical Evergreen installation.  They will also get
defaults from the PostgreSQL environment variables if these are set.

The file parameter defaults to parameters.xlsx and will create the
file in the current working directory if it is omitted.

Signed-off-by: Jason Stephenson <jason@sigio.com>